Chance buffsize from int to size_t

+++ monty.c
@@ -8,7 +8,7 @@
 
 int main(int argc, char *argv[])
 {
-    int buffsize = 512, data = 0, count = 0;
+    size_t buffsize = 512, data = 0, count = 0;
     char *buffer = malloc(buffsize * sizeof(char)), *temp = NULL, *cmd;
     stack_t *stack = NULL;
     FILE *file = fopen(argv[1], "r");
